chore: add test for custom zk chain (#105)
Adds a test to run the "custom zk chain" tutorial e2e locally. The test
took me ~35 minutes to run. I tried to add it as a separate workflow
that can be run manually, but I was having some issue with the
dependencies. I can follow up in another PR to try to get that working.

To test locally, make sure Docker is running in the background first.